requirements.txt
streamlit
numpy
pandas
scikit-learn
faiss-cpu
sentence-transformers
langchain
langchain-community
matplotlib
pypdf
reportlab
joblib
seaborn
python-dotenv
